some fixes added new skills qol changes

+++ src/ui/player/CharacterBag.h
@@ -76,7 +76,7 @@ void EntityStats::UnEquipItem(const float* effect_arr) noexcept {
   for (uint_fast32_t i = 0; i < STATS_ENDING; i++) {
     effects[i] -= effect_arr[i];
   }
-  CharacterBag::RemoveSlots(effect_arr[BAG_SLOTS]);
+  CharacterBag::RemoveSlots((int)effect_arr[BAG_SLOTS]);
   ReCalculatePlayerStats();
 }
